// src/resources/posting/interface.ts
export enum AwolSearchCriteria{
	PERSONNEL_NO="PERSONNEL_NO",
	EFFECTIVE_DATE_OF_AWOL="EFFECTIVE_DATE_OF_AWOL",
	AWOL_AUTHORITY="AWOL_AUTHORITY",
	AUTHORITY = "AUTHORITY",
	EFFECTIVE_DATE_OF_AWOL_CANCELLATION = "EFFECTIVE_DATE_OF_AWOL_CANCELLATION",
	AWOL_CANCELLATION_AUTHORITY = "AWOL_CANCELLATION_AUTHORITY",
 }

 

 export interface PostingType {
	SERIAL?: number;
	"PERSONNEL NO": string;
	"NAME": string;
	"UNIT FROM": string;
	"UNIT TO": string;
	APPOINTMENT: string;
	AUTHORITY: string;
	"EFFECTIVE DATE": string;
	REMARKS: string;
}